Implemented the logic to connect the repository logic with the retrofit API service.
Added the logic to manage the authentication with the API
Using LiveData objects to keep up-to-date the UI with the latest changes from the Service
Implemented a repository policy mechanism that declares the DB as the unique source of true; using the API service just as a repo to refresh the local data.
Added a network security configuration policy to debug the api requests with Charles.
Created a LoginUseCase to manage the login with the selected architecture as the rest of the app.
Added timber to manage the app logs

Submitted by Victor Valencia <cytrux@gmail.com>